Move store.js
