+++ app/javascript/components/PostPage.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Provider } from 'react-redux'
-import { store } from './store'
+import { store } from '../store'
 
 import PostEditor from './PostEditor'
 import ImageModal from './ImageModal'
