+++ app/javascript/store.js
@@ -1,4 +1,5 @@
-import { configureStore, createReducer, createSlice } from '@reduxjs/toolkit'
+import { configureStore, createReducer } from '@reduxjs/toolkit'
+import images from './reducers/imagesReducer'
 
 // unsafe side-effects for comment actions
 const commentsEffects = {
@@ -45,27 +46,6 @@ const commentReducers = {
     state.newestComment = action.payload
   },
 }
-
-const images = createSlice({
-  name: 'images',
-  initialState: {
-    isAddingimage: false,
-    lastImage: null,
-  },
-  reducers: {
-    addImageStart: (state, action) => {
-      state.lastImage = null
-      state.isAddingImage = true
-    },
-    addImageSuccess: (state, action) => {
-      state.lastImage = action.payload
-      state.isAddingImage = false
-    },
-    closeImageModal: (state, action) => {
-      state.isAddingImage = false
-    },
-  },
-})
 
 const postReducers = {
   setCurrentPost: (state, action) => {
