+++ package.json
@@ -35,6 +35,7 @@
     "date-fns": "^2.8.1",
     "es6-promise": "^4.2.8",
     "html-react-parser": "^0.13.0",
+    "immer": "^7.0.8",
     "jquery": "^3.4.1",
     "lodash": "4.17.19",
     "popper.js": "^1.16.0",
